move set_version_globals to utils
